+++ cypress/locators/flat-table/locators.js
@@ -3,3 +3,6 @@ export const FLAT_TABLE_COMPONENT = '[data-component="flat-table"]';
 export const FLAT_TABLE_STICKY_ROW = '[data-element="flat-table-row-header"]';
 export const FLAT_TABLE_CELL = '[data-element="flat-table-cell"]';
 export const FLAT_TABLE_SUBROW = '[data-element="flat-table-sub-row"]';
+export const FLAT_TABLE_PAGE_SIZE_SELECT = '[data-element="dropdown"]';
+export const FLAT_TABLE_PAGE_SELECT_LIST =
+  '[data-element="select-list-wrapper"]';