+++ cypress/support/step-definitions/select-steps.js
@@ -21,6 +21,7 @@ import {
   multiColumnsSelectListBody,
   boldedAndUnderlinedValue,
   selectInput,
+  selectListPosition,
 } from "../../locators/select";
 import { positionOfElement, keyCode } from "../helper";
 import { label, getDataElementByValue } from "../../locators";
@@ -242,4 +243,20 @@ Then("The matching string {string} is underline and bolded", (text) => {
 
 Then("I type {string} into select input", (text) => {
   selectInput().type(text);
+});
+
+When("I scroll page to top", () => {
+  cy.scrollTo("0", "0");
+});
+
+Then("{string} Select list is visible in {word} position", (name, position) => {
+  selectListPosition(name)
+    .should("have.attr", "data-popper-placement", position)
+    .and("be.visible");
+});
+
+Then("{string} Select list is visible at the {word}", (name, position) => {
+  selectListPosition(name)
+    .should("have.attr", "data-popper-placement", `${position}-end`)
+    .and("be.visible");
 });