+++ src/__experimental__/components/label/label.component.js
@@ -98,6 +98,7 @@ const Label = ({
             warning={warning}
             info={info}
             tooltipPosition={tooltipPositionValue}
+            tooltipFlipOverrides={["top", "bottom"]}
           />
         </IconWrapperStyle>
       );