+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+### [68.3.1](https://github.com/Sage/carbon/compare/v68.3.0...v68.3.1) (2021-03-15)
+
+
+### Bug Fixes
+
+* **menu:** fix search styles ([ce4405e](https://github.com/Sage/carbon/commit/ce4405ea5d8abbff2d032d756f0fdbb35b2835de)), closes [#3591](https://github.com/Sage/carbon/issues/3591)
+* **menu:** fix search styles ([0efe3dd](https://github.com/Sage/carbon/commit/0efe3dd07e82a9002c57985886a8e887e008b56d)), closes [#3591](https://github.com/Sage/carbon/issues/3591)
+
+## [68.3.0](https://github.com/Sage/carbon/compare/v68.2.0...v68.3.0) (2021-03-12)
+
+
+### Features
+
+* **link:** add isSkipLink prop ([cfefb6f](https://github.com/Sage/carbon/commit/cfefb6fd584e9aefb1abc4cdf4de84ed867bbe8a))
+
+## [68.2.0](https://github.com/Sage/carbon/compare/v68.1.1...v68.2.0) (2021-03-12)
+
+
+### Features
+
+* **button-toggle-group:** add support for overriding tooltip flip behaviour ([79d5f23](https://github.com/Sage/carbon/commit/79d5f239069fa771543802deb5a9a1b2f58ab6be))
+* **fieldset:** add tooltip flip override to validation icon rendered on labels ([2abfbd4](https://github.com/Sage/carbon/commit/2abfbd44528274f8363f1ee1c684f40fc76947e3))
+* **help:** add support for tooltip flip placement overrides ([2c01499](https://github.com/Sage/carbon/commit/2c014998f6166556d5e937ee8d5c3c4a37ae5267))
+* **icon:** add support for tooltip flip placement overrides ([47dc424](https://github.com/Sage/carbon/commit/47dc4244acdae1b53ca4ad4857803d51ab831bd9))
+* **label:** add tooltip flip override to validation icon when rendered ([2b88b25](https://github.com/Sage/carbon/commit/2b88b25242f5862584ff03b6aaa832537b8267ec))
+* **simple-color-picker:** add tooltip flip overrides to validation icon ([224d99d](https://github.com/Sage/carbon/commit/224d99d8747c03a5e558118a11a07771fb4e9257))
+* **switch:** add tooltip flip override to validation icon rendered on input ([ba6b73a](https://github.com/Sage/carbon/commit/ba6b73a074648f0558536962352b629b7727e2d8))
+* **tooltip:** add support for overriding default flip placements ([73a35d6](https://github.com/Sage/carbon/commit/73a35d61a360d8680e7703472cf5abe6f194acfd))
+* **validation-icon:** add support for tooltip flip placement overrides ([0471c5b](https://github.com/Sage/carbon/commit/0471c5b612a37f42045e6f7237559ee81b28f5ae))
+
+### [68.1.1](https://github.com/Sage/carbon/compare/v68.1.0...v68.1.1) (2021-03-11)
+
+
+### Bug Fixes
+
+* **select:** fix incorrect textbox behaviour on safari ([6077c18](https://github.com/Sage/carbon/commit/6077c186ede063693ea44fd04a33c3684a906d42))
+
+## [68.1.0](https://github.com/Sage/carbon/compare/v68.0.2...v68.1.0) (2021-03-11)
+
+
+### Features
+
+* **action-popover-item:** support for download button ([0dcdb65](https://github.com/Sage/carbon/commit/0dcdb654473e007d8a89787d99b9034f358c3350))
+
+### [68.0.2](https://github.com/Sage/carbon/compare/v68.0.1...v68.0.2) (2021-03-11)
+
+
+### Bug Fixes
+
+* **button:** add border radius to button styles ([3142e17](https://github.com/Sage/carbon/commit/3142e170fb9182c67d3749b1b36d667a49fab30a)), closes [#3630](https://github.com/Sage/carbon/issues/3630)
+
 ### [68.0.1](https://github.com/Sage/carbon/compare/v68.0.0...v68.0.1) (2021-03-09)